[feat] Add new python build script
The new python build script will now safely update the llvm repo and
build it only if needed, to safe time. It achieves this, by storing the
current git commit hash after building, so that it can check, if the
hash has changed, when it is run the next time.