[style][WIP] Change to LLVM-style error handling in the Parser class
LLVM recommends in its [documentation] to not use c++ exceptions, but
rather explicitly return `llvm::Error`s and `llvm::Expected`s. The Par-
ser class is the first of two places, where exceptions were used, and it
is now completely rewritten for the new error handling.
The code in this commit will still not compile.

[documentation]: http://llvm.org/docs/ProgrammersManual.html#recoverable-errors